# openapi/authenticationprovider-v1.yaml
openapi: 3.0.1
info:
  description: OpenAPI definition for the Authentication Provider component of the
    SIMPL project
  license:
    name: European Union Public License (EUPL) 1.2
    url: https://interoperable-europe.ec.europa.eu/sites/default/files/custom-page/attachment/eupl_v1.2_en.pdf
  title: Authentication Provider OpenAPI definition
  version: 1.3.0
servers:
- description: authenticationprovider
  url: /v1
...
paths:
  /keypairs:
    get:
      description: ONBOARDER_M user get the public/private Key Pair
      operationId: getInstalledKeyPair
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyPair'
          description: Key Pair successfully retrieved. The response body contains
            the details of the requested KeyPair.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unauthorized
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Key Pair not found. The requested KeyPair does not exist or
            is not accessible.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Conflict
      security:
      - oauth2ClientCredentials: []
      summary: Get installed Key Pair
      tags:
      - Keypairs
    head:
      description: ONBOARDER_M checks whether a public/private Key Pair is stored
        in the database
      operationId: existsKeypair
      responses:
        "200":
          description: KeyPair is present
        "401":
          description: Unauthorized
        "404":
          description: KeyPair is not present
      security:
      - oauth2AuthCode: []
      summary: Keypair Exists
      tags:
      - Keypairs
      x-simpl-roles:
      - ONBOARDER_M
    post:
      description: ONBOARDER_M user import a public/private Key Pair and store it
        to database
      operationId: importKeyPair
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportKeyPair'
        required: true
      responses:
        "204":
          description: Key Pair generate and stored successfully
        "401":
          description: Unauthorized
        "409":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Conflict
      security:
      - oauth2AuthCode: []
      summary: Import Key Pair
      tags:
      - Keypairs
      x-simpl-roles:
      - ONBOARDER_M
  /keypairs/generate:
    post:
      description: ONBOARDER_M user generate a public/private Key Pair and store it
        to database
      operationId: generateKeyPair
      responses:
        "204":
          description: Key Pair generate and stored successfully
        "401":
          description: Unauthorized
        "409":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Conflict
      security:
      - oauth2AuthCode: []
      summary: Generate Key Pair
      tags:
      - Keypairs
      x-simpl-roles:
      - ONBOARDER_M
...
  /credentials/publicKey:
    get:
      deprecated: true
      description: Fetches the public key associated with the stored credential
      operationId: getPublicKey
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credential'
          description: Successfully retrieved public key
        "404":
          description: Credential not found
      summary: Use GET /credentials
      tags:
      - Credentials
...
  /mtls/ephemeralProof:
    post:
      description: Stores the ephemeral proof for a participant identified by their
        UUID
      operationId: storeCallerEphemeralProof
      parameters:
      - description: The Public Key Hash of the participant
        explode: false
        in: header
        name: Credential-Id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          text/plain:
            schema:
              type: string
        description: The ephemeral proof to be stored
        required: true
      responses:
        "200":
          description: Ephemeral proof successfully stored
        "400":
          description: Invalid input data
        "404":
          description: Participant not found
      security:
      - mtls: []
      summary: Store Ephemeral Proof
      tags:
      - Mtls
  /sessions/credentials:
    post:
      description: Validate the tier one session against the ephemeral proof stored
        in the agent
      operationId: validateTierOneSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TierOneSession'
        required: true
      responses:
        "204":
          description: Tier 1 session validated successfully
        "422":
          description: Invalid Tier 1 session
      security:
      - oauth2ClientCredentials: []
      summary: Validate Tier 1 session
      tags:
      - Sessions
  /sessions/identityAttributes:
    get:
      description: Fetches the identity attributes associated with the specified participant
        ID
      operationId: getIdentityAttributesOfParticipant
      parameters:
      - description: The Public Key Hash of the participant
        explode: false
        in: query
        name: credentialId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/IdentityAttribute'
                type: array
          description: Successfully retrieved identity attributes
        "404":
          description: Ephemeral proof not found
      security:
      - oauth2ClientCredentials: []
      summary: Retrieve identity attributes of a participant
      tags:
      - Sessions
...
components:
  schemas:
    ProblemDetail:
      example:
        instance: https://openapi-generator.tech
        detail: detail
        type: https://openapi-generator.tech
        title: title
        properties:
          key: "{}"
        status: 0
      properties:
        type:
          format: uri
          type: string
        title:
          type: string
        status:
          format: int32
          type: integer
        detail:
          type: string
        instance:
          format: uri
          type: string
        properties:
          additionalProperties:
            type: object
          type: object
      type: object
    ImportKeyPair:
      example:
        privateKey: privateKey
        publicKey: publicKey
      properties:
        publicKey:
          type: string
        privateKey:
          type: string
      required:
      - privateKey
      - publicKey
      type: object
    KeyPair:
      example:
        privateKey: privateKey
        publicKey: publicKey
      properties:
        publicKey:
          format: byte
          type: string
        privateKey:
          format: byte
          type: string
      type: object
...
    IdentityAttribute:
      example:
        code: code
        name: name
        creationTimestamp: 2000-01-23T04:56:07.000+00:00
        description: description
        assignableToRoles: true
        participantTypes:
        - participantTypes
        - participantTypes
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        used: true
        enabled: true
        updateTimestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          format: uuid
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        assignableToRoles:
          type: boolean
        enabled:
          type: boolean
        creationTimestamp:
          format: date-time
          type: string
        updateTimestamp:
          format: date-time
          type: string
        participantTypes:
          items:
            type: string
          type: array
          uniqueItems: true
        used:
          readOnly: true
          type: boolean
      required:
      - assignableToRoles
      - code
      - enabled
      - name
      type: object
...
    Credential:
      example:
        participantId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        credentialId: credentialId
        publicKey: publicKey
      properties:
        publicKey:
          type: string
        credentialId:
          type: string
        participantId:
          format: uuid
          type: string
      type: object
...
    TierOneSession:
      example:
        jwt: jwt
      properties:
        jwt:
          type: string
      required:
      - jwt
      type: object
...
  securitySchemes:
    mtls:
      description: SIMPL mTLS
      scheme: http
      type: http
    oauth2AuthCode:
      flows:
        authorizationCode:
          authorizationUrl: ""
          scopes: {}
          tokenUrl: ""
      type: oauth2
    oauth2ClientCredentials:
      flows:
        clientCredentials:
          scopes: {}
          tokenUrl: ""
      type: oauth2